requirements.txt
# SPDX-License-Identifier: AGPL-3.0-or-later
# Copyright (C) 2026 Relational Network

# Azure SDK dependencies
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.1
azure-mgmt-network>=25.1.0
azure-mgmt-compute>=30.3.0

# FastAPI and web server
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.1

# Utilities
python-dotenv>=1.0.1
click>=8.1.7

# Testing
pytest>=8.0.1
httpx>=0.27.0

# Code quality
black>=24.2.0
flake8>=7.0.0